+++ Scenarios/Core.Services.Authenticate/Core.Services.Authenticate.WebAPI/Models/Item.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace Core.Services.Authenticate.WebAPI.Models
+{
+    public class Item
+    {
+        public int Id { get; set; }
+        public string Title { get; set; }
+    }
+}