+++ Solutions/PowerShell.Commands/Commands/Workflows/GetWorkflowDefinition.cs
@@ -1,0 +1,38 @@
+﻿using System.Management.Automation;
+using Microsoft.SharePoint.Client;
+using Microsoft.SharePoint.Client.WorkflowServices;
+using OfficeDevPnP.PowerShell.Commands.Base.PipeBinds;
+using System.Linq;
+
+namespace OfficeDevPnP.PowerShell.Commands.Workflows
+{
+    [Cmdlet(VerbsCommon.Get, "SPOWorkflowDefinition")]
+    public class GetWorkflowDefinition : SPOWebCmdlet
+    {
+        [Parameter(Mandatory = false, HelpMessage = "The name of the workflow", Position = 0)]
+        public string Name;
+
+        [Parameter(Mandatory = false)]
+        public SwitchParameter PublishedOnly = true;
+
+        protected override void ExecuteCmdlet()
+        {
+            if (string.IsNullOrEmpty(Name))
+            {
+                var servicesManager = new WorkflowServicesManager(ClientContext, this.SelectedWeb);
+                var deploymentService = servicesManager.GetWorkflowDeploymentService();
+                var definitions = deploymentService.EnumerateDefinitions(PublishedOnly);
+
+                ClientContext.Load(definitions);
+
+                ClientContext.ExecuteQuery();
+                WriteObject(definitions, true);
+            }
+            else
+            {
+                WriteObject(this.SelectedWeb.GetWorkflowDefinition(Name, PublishedOnly));
+            }
+        }
+    }
+
+}