+++ Solutions/PowerShell.Commands/changelog.md
@@ -1,5 +1,13 @@
 # OfficeDevPnP.PowerShell Changelog #
 
+**2014-10-15**
+* Added Add-SPOWorkflowSubscription, Get-SPOWorkflowDefinition, Get-SPOWorkflowSubscription, Remove-SPOWorkflowDefinition, Remove-SPOWorkflowSubscription, Resume-SPOWorkflowInstance, Stop-SPOWorkflowInstance
+
+**2014-10-14**
+* Added Get-SPOUserProfileProperty cmdlet
+* Added New-SPOPersonalSite cmdlet
+* Fixed Get-SPOView cmdlet
+
 **2014-10-08**
 * Added Set-SPODefaultColumnValue 
 **2014-09-19**